requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
imageio>=2.15.0
tqdm>=4.62.0
PyQt6>=6.2.0